Improvements: - Added TopLoader to manage loading - Minor interface fixes
