+++ app/(home)/page.tsx
@@ -7,19 +7,26 @@ import ThreatsCard from "@/components/threats-card";
 import { fetchData } from "@/lib/data-fetcher";
 import { useUser } from "@clerk/nextjs";
 import DeviceRadialChart from "@/components/device-radial-chart";
+import { useRouter } from "next/navigation";
+import TopLoader from "nextjs-toploader"; // Import TopLoader
 
 export default function Dashboard() {
   const [data, setData] = useState<any>(null);
-  const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
   const [privateMetadata, setPrivateMetadata] = useState<any>(null);
   const { isLoaded, isSignedIn, user } = useUser();
+  const router = useRouter();
 
+  // Handle authentication and redirect if necessary
   useEffect(() => {
-    if (!isLoaded || !isSignedIn) {
-      setLoading(false);
-      return;
+    if (isLoaded && !isSignedIn) {
+      router.push("/sign-in"); // Redirect to sign-in if not authenticated
     }
+  }, [isLoaded, isSignedIn, router]);
+
+  // Fetch private metadata after authentication is loaded
+  useEffect(() => {
+    if (!isLoaded || !isSignedIn || !user) return;
 
     const fetchPrivateMetadata = async () => {
       try {
@@ -33,14 +40,13 @@ export default function Dashboard() {
         }
       } catch (err: any) {
         setError("Error fetching metadata: " + err.message);
-      } finally {
-        setLoading(false);
       }
     };
 
     fetchPrivateMetadata();
   }, [isLoaded, isSignedIn, user]);
 
+  // Fetch data based on metadata
   useEffect(() => {
     if (!privateMetadata || !user) return;
 
@@ -60,21 +66,28 @@ export default function Dashboard() {
         setData(result);
       } catch (err: any) {
         setError(err.message);
-      } finally {
-        setLoading(false);
       }
     };
 
     getData();
   }, [privateMetadata, user]);
 
-  if (loading) return <p>Loading...</p>;
+  // Ensure the page doesn't render until authentication state is loaded
+  if (!isLoaded) {
+    return (
+      <div className="flex justify-center items-center h-screen">
+        <TopLoader />
+      </div>
+    );
+  }
+
   if (error) return <p>Internal Error: {error}</p>;
 
   // Calculate the active and offline devices
   const activeDevices = data ? data.count : 0;
   const totalDevices = data ? data.total : 0;
   const offlineDevices = totalDevices - activeDevices;
+
   return (
     <>
       <div className="flex flex-row items-center justify-center space-x-6 p-10">
